fix(bom): BOMComponentDetail unit_cost/extended_cost int -> float (handle Decimal)

+++ backend/api_gateway/app/schemas/bom.py
@@ -15,10 +15,8 @@ from pydantic import BaseModel, Field
 # WORK CENTERS
 # ============================================================================
 
-
 class CreateWorkCenterRequest(BaseModel):
     """Request to create work center"""
-
     code: str = Field(..., min_length=1, max_length=50)
     name: str = Field(..., min_length=1, max_length=100)
     description: Optional[str] = None
@@ -31,7 +29,6 @@ class CreateWorkCenterRequest(BaseModel):
 
 class UpdateWorkCenterRequest(BaseModel):
     """Request to update work center"""
-
     name: Optional[str] = Field(None, min_length=1, max_length=100)
     description: Optional[str] = None
     warehouse_id: Optional[UUID] = None
@@ -44,7 +41,6 @@ class UpdateWorkCenterRequest(BaseModel):
 
 class WorkCenterListItem(BaseModel):
     """Work center in list view"""
-
     id: str
     code: str
     name: str
@@ -56,7 +52,6 @@ class WorkCenterListItem(BaseModel):
 
 class WorkCenterListResponse(BaseModel):
     """Response for listing work centers"""
-
     items: List[WorkCenterListItem]
     total: int
     has_more: bool
@@ -64,7 +59,6 @@ class WorkCenterListResponse(BaseModel):
 
 class WorkCenterDetail(BaseModel):
     """Work center detail"""
-
     id: str
     code: str
     name: str
@@ -82,7 +76,6 @@ class WorkCenterDetail(BaseModel):
 
 class WorkCenterDetailResponse(BaseModel):
     """Response for work center detail"""
-
     success: bool = True
     data: WorkCenterDetail
 
@@ -91,10 +84,8 @@ class WorkCenterDetailResponse(BaseModel):
 # BOM COMPONENTS
 # ============================================================================
 
-
 class BOMComponentInput(BaseModel):
     """Input for BOM component"""
-
     component_product_id: UUID
     quantity: Decimal = Field(..., gt=0)
     unit: Optional[str] = None
@@ -107,7 +98,6 @@ class BOMComponentInput(BaseModel):
 
 class BOMComponentDetail(BaseModel):
     """BOM component detail"""
-
     id: str
     component_product_id: str
     component_product_name: str
@@ -120,8 +110,8 @@ class BOMComponentDetail(BaseModel):
     wastage_percent: Decimal
     operation_id: Optional[str]
     operation_name: Optional[str]
-    unit_cost: int
-    extended_cost: int
+    unit_cost: float
+    extended_cost: float
     notes: Optional[str]
     sequence_order: int
     is_substitute: bool
@@ -132,12 +122,11 @@ class BOMComponentDetail(BaseModel):
 # BOM OPERATIONS
 # ============================================================================
 
-LaborMode = Literal["time_based", "piece_rate", "subcontract", "none"]
+LaborMode = Literal['time_based', 'piece_rate', 'subcontract', 'none']
 
 
 class BOMOperationInput(BaseModel):
     """Input for BOM operation"""
-
     operation_number: int = Field(..., ge=1)
     operation_name: str = Field(..., min_length=1, max_length=100)
     description: Optional[str] = None
@@ -153,13 +142,12 @@ class BOMOperationInput(BaseModel):
     subcontract_cost_per_unit: float = 0
     subcontract_description: Optional[str] = None
     # V133 new
-    labor_mode: LaborMode = "time_based"
+    labor_mode: LaborMode = 'time_based'
     cost_per_piece: Optional[float] = None
 
 
 class BOMOperationDetail(BaseModel):
     """BOM operation detail"""
-
     id: str
     operation_number: int
     operation_name: str
@@ -177,7 +165,7 @@ class BOMOperationDetail(BaseModel):
     subcontract_cost_per_unit: float = 0
     subcontract_description: Optional[str] = None
     # V133 new
-    labor_mode: LaborMode = "time_based"
+    labor_mode: LaborMode = 'time_based'
     cost_per_piece: Optional[float] = None
 
 
@@ -185,10 +173,8 @@ class BOMOperationDetail(BaseModel):
 # BILL OF MATERIALS
 # ============================================================================
 
-
 class CreateBOMRequest(BaseModel):
     """Request to create BOM"""
-
     product_id: UUID
     bom_code: str = Field(..., min_length=1, max_length=50)
     bom_name: Optional[str] = Field(None, max_length=255)
@@ -204,7 +190,6 @@ class CreateBOMRequest(BaseModel):
 
 class UpdateBOMRequest(BaseModel):
     """Request to update BOM"""
-
     bom_name: Optional[str] = Field(None, max_length=255)
     description: Optional[str] = None
     output_quantity: Optional[Decimal] = Field(None, gt=0)
@@ -220,7 +205,6 @@ class UpdateBOMRequest(BaseModel):
 
 class BOMListItem(BaseModel):
     """BOM in list view"""
-
     id: str
     bom_code: str
     bom_name: Optional[str]
@@ -238,7 +222,6 @@ class BOMListItem(BaseModel):
 
 class BOMListResponse(BaseModel):
     """Response for listing BOMs"""
-
     items: List[BOMListItem]
     total: int
     has_more: bool
@@ -246,7 +229,6 @@ class BOMListResponse(BaseModel):
 
 class BOMDetail(BaseModel):
     """Detailed BOM with components and operations"""
-
     id: str
     bom_code: str
     bom_name: Optional[str]
@@ -276,7 +258,6 @@ class BOMDetail(BaseModel):
 
 class BOMDetailResponse(BaseModel):
     """Response for BOM detail"""
-
     success: bool = True
     data: BOMDetail
 
@@ -285,10 +266,8 @@ class BOMDetailResponse(BaseModel):
 # COST BREAKDOWN
 # ============================================================================
 
-
 class CostBreakdownItem(BaseModel):
     """Cost breakdown item"""
-
     category: str  # material, labor, overhead, subcontract
     description: str
     quantity: Optional[Decimal]
@@ -299,7 +278,6 @@ class CostBreakdownItem(BaseModel):
 
 class CostBreakdownResponse(BaseModel):
     """Response for cost breakdown"""
-
     success: bool = True
     bom_code: str
     product_name: str
@@ -317,10 +295,8 @@ class CostBreakdownResponse(BaseModel):
 # MATERIALS REQUIRED
 # ============================================================================
 
-
 class MaterialRequiredItem(BaseModel):
     """Material required for production"""
-
     product_id: str
     product_name: str
     product_sku: Optional[str]
@@ -334,7 +310,6 @@ class MaterialRequiredItem(BaseModel):
 
 class MaterialsRequiredResponse(BaseModel):
     """Response for materials required"""
-
     success: bool = True
     bom_code: str
     production_quantity: Decimal
@@ -347,10 +322,8 @@ class MaterialsRequiredResponse(BaseModel):
 # BOM EXPLOSION
 # ============================================================================
 
-
 class BOMExplosionNode(BaseModel):
     """Node in BOM explosion tree"""
-
     level: int
     product_id: str
     product_name: str
@@ -366,7 +339,6 @@ class BOMExplosionNode(BaseModel):
 
 class BOMExplosionResponse(BaseModel):
     """Response for BOM explosion (multi-level)"""
-
     success: bool = True
     bom_code: str
     product_name: str
@@ -379,10 +351,8 @@ class BOMExplosionResponse(BaseModel):
 # WHERE USED
 # ============================================================================
 
-
 class WhereUsedItem(BaseModel):
     """Item showing where component is used"""
-
     bom_id: str
     bom_code: str
     bom_name: Optional[str]
@@ -395,7 +365,6 @@ class WhereUsedItem(BaseModel):
 
 class WhereUsedResponse(BaseModel):
     """Response for where-used query"""
-
     success: bool = True
     product_id: str
     product_name: str
@@ -407,10 +376,8 @@ class WhereUsedResponse(BaseModel):
 # GENERIC RESPONSES
 # ============================================================================
 
-
 class BOMResponse(BaseModel):
     """Generic response for BOM operations"""
-
     success: bool
     message: str
     data: Optional[Dict[str, Any]] = None
